break appart the task stuff
