+++ raphtory/src/db/task/context.rs
@@ -1,0 +1,122 @@
+use std::sync::Arc;
+
+use crate::{
+    core::{
+        agg::Accumulator,
+        state::{AccId, ComputeState, ShuffleComputeState, StateType},
+    },
+    db::view_api::internal::GraphViewInternalOps,
+};
+
+type MergeFn<CS> =
+    Arc<dyn Fn(&mut ShuffleComputeState<CS>, &ShuffleComputeState<CS>, usize) + Send + Sync>;
+
+pub struct Context<G, CS>
+where
+    G: GraphViewInternalOps + Send + Sync + 'static,
+    CS: ComputeState,
+{
+    ss: usize,
+    g: Arc<G>,
+    merge_fns: Vec<MergeFn<CS>>,
+    resetable_states: Vec<u32>,
+}
+
+impl<G, CS> Context<G, CS>
+where
+    G: GraphViewInternalOps + Send + Sync + 'static,
+    CS: ComputeState,
+{
+    pub fn ss(&self) -> usize {
+        self.ss
+    }
+
+    pub fn graph(&self) -> Arc<G> {
+        self.g.clone()
+    }
+
+    pub fn increment_ss(&mut self) {
+        self.ss += 1;
+    }
+
+    pub fn resetable_states(&self) -> &[u32] {
+        &self.resetable_states
+    }
+
+    pub(crate) fn run_merge(
+        &self,
+        mut a: Arc<ShuffleComputeState<CS>>,
+        mut b: Arc<ShuffleComputeState<CS>>,
+    ) -> Arc<ShuffleComputeState<CS>> {
+        if let Some(left) = Arc::get_mut(&mut a) {
+            for merge_fn in self.merge_fns.iter() {
+                merge_fn(left, &b, self.ss);
+            }
+            a
+        } else if let Some(right) = Arc::get_mut(&mut b) {
+            for merge_fn in self.merge_fns.iter() {
+                merge_fn(right, &a, self.ss);
+            }
+            b
+        } else {
+            // none of the states have been changes so just return one of them
+            a
+        }
+    }
+
+    pub fn agg<A: StateType, IN: 'static, OUT: 'static, ACC: Accumulator<A, IN, OUT>>(
+        &mut self,
+        id: AccId<A, IN, OUT, ACC>,
+    ) {
+        let fn_merge: MergeFn<CS> = Arc::new(move |a, b, ss| a.merge_mut_2(b, id, ss));
+
+        self.merge_fns.push(fn_merge);
+    }
+
+    pub fn agg_reset<A: StateType, IN: 'static, OUT: 'static, ACC: Accumulator<A, IN, OUT>>(
+        &mut self,
+        id: AccId<A, IN, OUT, ACC>,
+    ) {
+        let fn_merge: MergeFn<CS> = Arc::new(move |a, b, ss| a.merge_mut_2(b, id, ss));
+
+        self.merge_fns.push(fn_merge);
+        self.resetable_states.push(id.id());
+    }
+
+    pub fn global_agg<A: StateType, IN: 'static, OUT: 'static, ACC: Accumulator<A, IN, OUT>>(
+        &mut self,
+        id: AccId<A, IN, OUT, ACC>,
+    ) {
+        let fn_merge: MergeFn<CS> = Arc::new(move |a, b, ss| a.merge_mut_global(b, id, ss));
+
+        self.merge_fns.push(fn_merge);
+    }
+
+    pub fn global_agg_reset<
+        A: StateType,
+        IN: 'static,
+        OUT: 'static,
+        ACC: Accumulator<A, IN, OUT>,
+    >(
+        &mut self,
+        id: AccId<A, IN, OUT, ACC>,
+    ) {
+        let fn_merge: MergeFn<CS> = Arc::new(move |a, b, ss| a.merge_mut_global(b, id, ss));
+
+        self.merge_fns.push(fn_merge);
+        self.resetable_states.push(id.id());
+    }
+}
+
+impl<G: GraphViewInternalOps + Send + Sync + Clone + 'static, CS: ComputeState> From<&G>
+    for Context<G, CS>
+{
+    fn from(g: &G) -> Self {
+        Self {
+            ss: 0,
+            g: Arc::new(g.clone()),
+            merge_fns: vec![],
+            resetable_states: vec![],
+        }
+    }
+}
